Move System.Runtime.InteropServices.Variant to shared directory
Reconcile Variant used for dynamic support with shared source
