+++ src/libraries/Common/src/Interop/Windows/OleAut32/Interop.VariantClear.cs
